requirements.txt
google-generativeai
phi
phidata
duckduckgo-search
streamlit
duckdb
python-dotenv
psycopg2
pandas
prophet
googlesearch-python
pycountry